src/part2/services/visual_indexer/requirements.txt
kafka-python>=2.0
byaldi>=0.0.6
colpali-engine>=0.3.8
transformers>=4.47.0,<4.48.0
bitsandbytes>=0.43.0
